Added Del's patch for: https://bugs.eclipse.org/bugs/show_bug.cgi?id=151889
Fixes:
Fixes coupling problems between NestedGraphViewerImpl, tree viewers, and the
BreadCrumbBar

NestedGraphViewer and NestedGraphViewerImpl no longer contain a TreeViewer or a
BreadCrumbBar. To allow NestedGraphViewer to interact with such other viewers,
however, the following functionality has been added:

1) NestedGraphViewer.add/removeFocusListener(INestedGraphFocusChangedListener)
has been added to support listeners that would like to know when an element has
been focussed on (i.e. zoomed on) in the viewer.

2)get/setFocus() has been exposed in NestedGraphViewer to allow focus to be
programatically set on the viewer.

3)Navigation functionality: (can)goBackward(), (can)goForward(), goUp(), and
focusHasParent(), have been created on NestedGraphViewer.

4) New property: GraphModel.NODE_FOCUS_PROP has been added so that viewers can
listen to the model to find out when focus has changed in the model.

5) new class BreadCrumbHook has been added to hook a BreadCrumbBar to a
NestedGraphViewer. BreadCrumbBars can exist without a NestedGraphViewer, and
vice-versa.