[162111] gef-head anyssen 101027 AbstractEditPart.getRoot() should not throw NPE if parent is null
Changed implementation of getRoot() and getViewer() in AbstractEditPart to not throw an NPE if called in case no parent is set. Changed JavaDoc of respective methods in EditPart to state that the methods may return a null value. Also changed the JavaDoc of getModel() to state that the method may return null as well.