[330859] gef-head anyssen 101129 SelectionTool, MarqueeSelectionTool and GraphicalViewerKeyHandler should use EditPart#isSelectable() consistently. "Invisible" GraphicalEditParts should not be selectable (or focusable).
Changed implementation of AbstractEditPart#isSelectable() and AbstractGraphicalEditPart#isSelectable() so that only active edit parts are selectable, and furthermore graphical edit parts have to have a showing figure. Clarified that isSelectable() is not only the precondition for edit parts to get selected, but also to obtain focus (as focus is considered part of the selection state). Added assertions to AbstractEditPart#setFocus(boolean) and AbstractEditPart#setSelected(int) to ensure the receiver is selectable in case focus is gained or the part is to be selected. Adjusted javadoc comments for setFocus(boolean), setSelected(int), and isSelectable().