[297223] gef-HEAD anyssen 110116 Shape#getLineStyle() overrides LineAttributes
Introduced a private lastLineStyle field to keep track of changes to the protected lineStyle field, and implemented similar mechanism as for lineWidth based on it. Deprecated lineWidth and lineStyle fields, because clients should rather use the setters or access the line attributes. Added a test case to assert the line style no longer gets overwritten.