[412089] Fix that TextStyle properties and font were lost during zoom.
- Fixed that within zoomTextLayout(TextLayou), strikeoutColor,
underlineColor, and underlineStyle of the original TextLayout's styles
were not transferred to the corresponding TextStyles of the newly
created zoomed TextLayout.
- Fixed that getCachedFontData(Font) returned the FontData of a passed
in font only, if there was already cached data present for it. In all
other cases the FontData of the local font (set on the Graphics) was
returned instead. As it may happen that there is no cached font data
present for Fonts being used in TextStyles, this was the cause for
loosing font height information when calculating TextStyles for the
newly created zoomed TextLayout inside zoomTextLayout(TextLayout); in
fact the Font set in the TextStyle was ignored and the local font of the
graphics was used to compute the zoomedFont instead.
- As zooming a TextLayout needs to include zooming of font height, a
font needs to be inferred, even if no Font is acutally set on the
original TextLayout, from which a zoomed one is created. Here,
zoomFont(Font) always used the display system fond to retrieve FontData
(which as a side effect returned the FontData of the local font). As the
behavior within getCachedFontData(Font) was fixed to no longer ignore
the passed in font (in case no cached data is present), ensured that
within zoomFont(Font), the local font gets precedence over the display
default font. This way, the usage of fonts within zooming of a
TextLayout is as follows:
	1) if a TextStyle provides a font, its FontData is used to compute a
new zoomed FontData
	2) if the TextStyle does not provide a font, but there is one set on
the Graphics (local font), this is used instead to compute FontData.
	3) if no font is provided by the TextStyle and no font is provided by
the Graphics (local font), then the display system font is taken as a
fallback