[323656] Add setHideDelay(int), ensured hide() cancels hide timer.
- Added a public setter by which the formerly hard-coded hide delay can
now be set. It is stored in a field and used within displayToolTipNear
as before. 
- To enable that subclasses may hide the tooltip before the hide timer
has actually expired, overwrote hide() to now cancel the hide timer as a
first step. Before, calls to hide() were always preceded or succeeded by
respective calls to Timer#cancel(). As it seems that in all these
situations canceling the timer first and hiding the figure afterwards is
proper, all these calls were removed.