[309912] Prevent figure can corrupt already painted areas in thumbnail.
- Added a tile image, into which the source figure now paints (instead
of painting directly into the thumbnail image). The thumnail image is
now udpated by copying the contents of this tile image into the
thumbnail (at the specified tile area) in each run. This way, even if
the source figure sets an own clip on the passed in Graphics inside
paint(Graphics), already painted areas in the thumbnail image can never
be corrupted (as we can control what gets copied).
- Fixed that the SWTGraphics that is constructed around the GC was never
disposed.