[332351] Add canRedo() to Command, evaluated by CommandStack#redo().
- Added canRedo() to Command, which by default delegates to
canExecute(), similar to redo() delegates to execute(). Subclasses may
overwrite to decide whether a command is redoable.
- Added an implementation for canRedo() within CompoundCommand, which
overwrites the default implementation within Command and delegates to
the canRedo() of the nested commands (similar as for canExecute() and
canUndo()).
- Changed implementation of CommandStack#canRedo() to evaluate the
canRedo() of the current redoable command.
- Fixed that undo() did not check that the undoable command is indeed
undoable (like done for redo()) but execute the undo in each case. Now,
as redo() it is a no-op in case the current undoable command is not
undoable.