Typefied exclusion search
With this cleanup the figure and edipart search was typified. As part of
this work it was identified that the TargetingTool documentation was
wrong. While there it was written that getExclusion set provides
editparts all its children provided a list of IFigures. Also the search
expected IFigures.

Furthermore AbstractTransferDropTargetListener delivered EditParts but
the search required IFigures. The code was fixed here was well and by
typifying the Collections it is now clearer.