[WB] Integrate geometry extensions from the WindowBuilder project.
The Draw2D fork used by the WindowBuilder project includes several
convenience methods. Because we want to migrate to the official GEF
release, those changes need to be merged back.

In addition to several convenience methods, a new Interval class is
introduced, describing the interval within 1-dimensional space. This
class is used by the Rectangle, in order to determine the width or the
height of a widget, depending on its orientation.

Test cases have been extended to cover this new functionality.