[WB] Create a new Interval class to describe distances in 1-d space.
The Interval is the counterpart to the Rectangle, for 1-dimensional
space. Its primary use is to describe the length of rows and columns of
layouts, in most cases the GridLayout.

The original class was created back in 2011 as part of the WindowBuilder
project and is now integrated into the GEF.