[WB] Add methods to modify the size and location of Rectangles.
Instead of only being able to shrink a Rectangle symmetrically (meaning
top/bottom and left/right), it should also be possible to shrink the top
or bottom side, only.

The new methods shrinkTop and shrinkBottom have been created. Each
method accepts a single argument, indicating by what amount the x or y
coordinate should be shifted.

In addition, the methods setBottom and setRight have been created, which
allow the user to alter the width or height of the rectangle. Each
method expects the absolute x/y coordinate of the right or bottom side
of the rectangle. The width/height is then adjusted to match those new
coordinates.