[JUnit] Stabilize tests on MacOS and define 5min timeout
It seems like the way timed jobs are executed on MacOS slightly differs
from Windows and Linux. As a result, it might happen that a job is
scheduled but the UI thread is never woken up again. As a result, the
test case is stuck indefinitely, until it is aborted by GitHub.

To avoid this problem, add this job asynchronously. This way, the job is
immediately added to the event queue. Furthermore, don't call sleep()
while pumping the event queue, in order to avoid missing out on wake-up
events.

To avoid issues like this in the future, specify a 5min timeout for each
test bundle. Given that we share our build nodes with the other Eclipse
projects, a job being stuck for 6h means that other developers might not
be able to run their builds for this amount of time. So we should fail
early, to free up those resources again.