[JUnit] Stabilize ConnectionEndPointMoveTest via waitEventLoop method
Use the same approach used to stabilize the PaintDamageErase tests. When
pumping the SWT event loop only once, the update of the UpdateManager
may have not been fully performed. In order to ensure that it has been
fully processed, pump the events for an arbitrary but fixed amount of
time.

For the sake of reusability, the waitEventLoop method has been moved to
the BaseTestCase class. Both test cases extend this base class in order
to access this method. The ConnectionEndPointMove test will pump SWT
events for a total of 100ms.