Adjust shrink method of rectangles to prevent negative width/height.
This negative value occurs when the rectangle has a border, but its
height or width is smaller than the height or width of its insets. The
shrink method then tries to reduce the width/height of the rectangle by
the width/height of the Insets which may then load to an error.