Fix: PointList bounds not updated when using setPoint()
The getBounds() method is supposed to return the minimal rectangle
enclosing all points.
When a point inside this PointList is updated via setPoint() and the
point is contained in the current bounds, then the current rectangle has
to be invalidated, s.t. getBounds() has to calculate a new, potentially
smaller rectangle.