Typify dirtyMap in UpdateListener and UpdateManager & cleanup usage
This change adjusts the notifyPainting() method in UpdateListener and
the firePainting() method in UpdateManager to expect a generic map of
(IFigure, Rectangle) pairs, rather than a raw map.

This should not cause errors in client code, as calling or overwriting
this method can still be done using a raw map (which will create a
"unchecked or unsafe operations" warning).

Internal usage has been changed from iterators to the for-each paradigm.