Fix: Regressions caused by cleaning up BreakCycle & PopulateRanks
The for-each loop inside the BreakCycle method used to iterate over all
graph nodes. This was wrongfully changed to a local variable which is
initialized as an empty list.

The for-each loop inside PopulateRanks iterates over all graph nodes and
calls VirtualNodeCreation internally. The constructor of this class,
however, may add nodes to the graph, leading to a
ConcurrentModifcationException. To avoid this problem, we iterate of a
local copy of the graph nodes instead.

Amends 31a70127f763d40fd1f6eb6a56de9158cf9538ce