Typify GroupRequest and SimpleFactory classes
For the GroupRequest class, we store the list of edit-parts as a
List<? extends EditPart>, in order to support casting it to e.g. a
List<CustomEditPart>.

For the SimpleFactory class, we replace the deprecated
Class.newInstance() call with Constructor.newInstance(), to keep the
compile-time exception checking (see JavaDoc for more details).