Create system cursors using Display getSystemCursor method #286 (#288)
Creating the static standard platform cursors may lead to a potential
memory leak, if the Cursors class ever happens to be unloaded, before
the application is terminated (e.g. when the bundle is stopped), because
this process won't free up the allocated resources.

To avoid this issue, we use Display.getSystemCursor, similar to what is
done to get the standard platform colors.

Given that those factory methods are technically API, they have also
been documented, with a large part taken from the getSystemCursor() and
getSystemColor() JavaDoc.