Specify different group id for features to avoid collisions with bundles
Both the GEF feature and the GEF bundle have the artifact id
"org.eclipse.gef".
If both artifact have the same group id, they are considered to be the
"same" artifact, from a Maven perspective.

Overwrite the group id for features to be "org.eclipse.gef.features", to
make sure the generated Maven coordinates are distinct.

See #290