Typify MultiValueMap and fix warnings in AutomaticRouter (#320)
The MultiValueMap used to be backed by a map with either a list or a
single object as value.

This part has been adapted so that the map always contains a list of
values. While this adds a small overhead when filling the map, it should
in return reduce the overhead when calling the getter for values with
only a single element.

Reduces the total amount of warnings by ~20.