Make default hide delay for tooltip configurable (#364)
The hide delay for tooltips in Draw2d is currently fixed to 5 seconds.
The value can only be changed for a speific `ToolTipHelper`, which in
turn is tied to a specific control. There is option to configure this
delay globally.

With this change, the default value for the hide delay can globally be
changed on the `ToolTipHelper` class, such that subsequent tooltip
creations will use this value as default.