Convert Help sources from HTML to AsciiDoctor (#351)
The documentation for the Draw2D, GEF and Zest bundles has been
converted to AsciiDoctor. The corresponding HTML files are generated as
part of the incremental build.

Meaning that any changes to the adoc files inside the IDE will cause the
HTML files to be updated allowing us to immediately verify the new
output.

Note that this change also performs minor changes to the documentation,
such as fixing URLs and references to the JavaDoc, as well as fixing the
JavaDoc generation for Zest itself.