Use optimized scrolling for all three scroll operations of FigureCanvas
FigureCanvas' scrollToX and scrollToY uses an optimized scrolling
functionality which leads especially for views with many figures to a
much smoother scrolling experience.

This commit utilizes this optimized scrolling also for scrollTo.

Furthermore the code for all three has been combined in such a way that
all three use the same core functions.