Handle implicit ClassCastException in Zest Core due to raw types
The selection of an abstract viewer can either be of type GraphNode or
GraphConnection. However, this list of items is explicitly casted to an
array of GraphNodes. Use the common base class GraphItem instead.

Furthermore, the viewer attempts to remove the objects of type
GraphConnection from the layout manager, whenever a node is removed. But
the layout manager only accepts objects of type LayoutRelationship which
is a member of the GraphConnection.