Add static method to configure overlay scrolling for FigureCanvas
This allows the user to specify whether new instances of FigureCanvas
should use overlay scrolling on not. This is disabled by default, due to
performance issues on Linux/GTK3.

As a result, the scrollbar is shown as a separate entity and no longer
part of the client area. This setting only affects operating systems
supporting overlay scrolling. This method requires SWT 3.126 or newer.

https://github.com/eclipse/gef-classic/issues/430