Move up null check in GEF viewer when the selected editpart is updated.
The selection inside a GEF viewer is assumed to always be non-null. But
because the (implicit) null check is done after the editpart is added to
the internal list, it leaves the viewer in an inconsistent state.

To avoid this, the null check is done earlier, so that the exception is
thrown before the internal model is modified.

Resolves https://github.com/eclipse/gef-classic/issues/433