Fix: Adding a node to a graph container adds it to the graph instead
When the constructors in the GraphNode class have been unified, we
wrongfully pass "graphModel.getGraph()" as an argument instead of the
current container. This causes all nodes to be added to the parent
graph.

See the UML example for reference. One of the class diagrams should be
contained by the UML container, but all three are part of the parent
graph instead.

Amends a13541fd21b23c8fdfb5c64a99321f7b8be44706