Integrate Zest-based examples into our test suite
With this change, we instantiate and verify the correctness of the Zest
examples as part of our JUnit test suite. Special care needs to be
taken, in order to make sure that the examples can be used in the test
environment, but also still be executed as standalone applications.

This means that those snippets must not create a new Display instance,
as one is already created by the test suite and multiple displays are
forbidden. The snippet is then instantiated by reflectively calling the
main method.

Because the snippets block until the shell is closed by the user, a
little trick is needed to allow for the tests to be executed. The actual
test is scheduled as an asynchronous task before the snippet is created,
which is then executed by the call to "Display.readAndDispatch()".

For this exact reason has the FileDialog been removed from
GraphJFaceSnippet3, as this would block the UI thread prematurely, thus
making testing impossible. The example is still functional, as it will
instead load the example file from the classpath.

In almost all cases, the SpringLayoutAlgorithm of the code snippets has
been replaced by a GridLayout, in order to make the test more stable and
reproducible.

This contribution is done in preparation for the Zest 2.0 migration as a
means to increase the test coverage and thus spot any changes that are
introduced.