Fix ConcurrentModificationException in RulerProvider
Use a thread-safe CopyOnWriteArrayList instead of a plain ArrayList for
storing all registered listeners. Note that writing to this type is
expensive, as the array is copied on every add. But for this use case,
the number of listeners should be negligible.

To reproduce: Enable the ruler in the Logic example and create a new
guide.