Fix: ConcurrentModificationException in CommandStack
Use a CopyOnWriteArrayList instead of a plain ArrayList for storing the
registered listeners. Adding or removing from this list always results
in a fresh copy of the underlying array, avoiding concurrent
modification.

Assuming that the number of times an event is fired far outweighs the
times a listener is added and/or removed, this should be more efficient
than creating a copy of the listeners, whenever the notify method is
called.

Resolves https://github.com/eclipse/gef-classic/issues/454