Fix: Marquee Selection Tool causes unresponsive UI
This commit partially reverts d94c624c8cd5b92b0568ac416bc400d68fe0ed7a.
When e.g. moving the mouse cursor out and back in the editor, the tool
is not properly reactivated.

Reason being that this only happens when the currently selected editor
is graphical viewer. But because the editor is still null (as it is set
to null when leaving the editor), this check always returns false.

The change was seemingly done as an optimization and is not relevant for
the actual commit. So it should be save to rroll it back.

Resolves https://github.com/eclipse/gef-classic/issues/466