Fix errors in drag/shared cursors when using high-DPI screens
The cursors used in the flyout palette only exist for the default 100%.
If a system is using a higher zoom level, the creation of those cursors
fails with an IllegalArgumentException. Instead, we should scale the
cursor image to match the current level.

As opposed to reading the zoom level only once, it is now calculated
dynamically, as the zoom may change depending on the display or because
it has been changed. The methods for handling this case have also been
moved from the SharedCursors class to the InternalGEFPlugin, so that it
becomes available to the DragCursors class.

From there, we call scaledImageData() to convert the image
descriptor to image data matching the requested zoom. The scaling is
done internally by SWT.

See https://github.com/eclipse/gef-classic/issues/464