Create SVG of Draw2D CheckBox icons and recreate them as png
This adds 96 and 192 DPI variants for the "checked" and "unchecked"
icons used in the CheckBox figure. A minor change has been done do the
images, where the white pixels outside of the check box have been
removed.

On Linux, the icons can be generated automatically via the build.sh
script in the "icons" folder, which also moves them to the correct
location in the bundle projects.