Convert back arrow of the Zest viewer to PNG and add high-DPI version
This adds a 96 and 192DPI version of the back arrow, as well as the
source SVG file from which they were created. In order to dynamically
update the images, the Draw2D FileImageDataProvider is used, rather than
loading the image directly from the classpath.

The image itself has been converted into a static variable, in order to
avoid a resource leak when the back button is removed from the graph
again.

See https://github.com/eclipse/gef-classic/issues/465