Fix: ViewportAutoexposeHelper is very intensive on CPU
The calls do doAutoexpose() are currently fired relentlessly one after
the other, leading to high (and mostly unnecessary) CPU usage. By using
the JFace Throttler class we add a small delay between those events.

The delay is an arbitrary but fixed value; Large enough to avoid the
excessive CPU cost, but small enough to not be noticeable by the user.

Resolves https://github.com/eclipse/gef-classic/issues/492