[Zest 2.0] Resolve Zest 1.0-related compiler warnings
Primarily related to the references of deprecated classes marked as
for-removal as well as the usage of raw types in the Zest 1.0 API. Where
possible, the warnings have been resolved properly as best as possibles.
But whenever this touches public API, we simply suppress the warning.

Given that this API will be removed in an upcoming release, the focus
should be on stability and to avoid any incompatibilities that might
arise from using the actual rather than raw types.