Duplicate palette icon to avoid issues when using the Palette View
The path of the icon URI in the plugin.xml file is using the workspace
path beginning with "src/...". But in the actual bundle, this prefix
doesn't exist. This commit restore the dedicated "icons" folder which
also holds an instance of this file, which was removed with
24daefbe8edfb9eb2e4c3db7b8f5e318b1b33860.

See https://github.com/eclipse/gef-classic/issues/512