[Zest 2.0] Fix: Deprecation warnings in GraphNode constructor
The constructors where one could pass the data object has been marked
for removal as part of the Zest 2.0 migration, with the suggestion to
call setData() afterwards.

However, this might not be possible for e.g. the CGraphNode, where the
data object is returned by a call to createFigureForModel(), which then
leads to an exception if this field hasn't been initialized yet.

To remedy this situation, the constructors expecting a raw object as
data remain deprecated and instead a new constructor has been created,
where this object has to be of type IFigure.