[Zest 2.0] Fix ambiguity in RotateGestureListener regarding connections
The RotateGestureListener assumes that it is only possible to select
nodes (and containers) inside a Zest graph, but not connections. Trying
to do so causes a ClassCastException.

To avoid this problem, we now explicitly check the currently selected
items and only perform the rotation if all of them are nodes.