Fix: ToolTipHelper doesn't work properly on Linux
The pop-up helper reuses the shell that is used to display the tooltip.
When a new tooltip is shown, the (currently invisible) shell is moved
from its old location and then made visible. This doesn't work in GTK
and the shell remains at its old location.

After the shell has been made visible, it is checked whether the shell
has been correctly moved and if necessary, the shell coordinates are
updated again. Note that it is still necessary to set the bounds while
the shell is invisible, in order to properly handle the first tooltip
that is shown.

Resolves https://github.com/eclipse/gef-classic/issues/531