[Zest 2.0] Resolve compiler warnings in Graph & friends
The IContainer2 interface expects the GraphContainer and Graph class to
return the package-private InternalLayoutContext when calling
getLayoutContext().

This doesn't work if this interface is used outside the Zest package
(e.g. in our tests) and prevents its from being implemented by clients.
Instead, we return the public LayoutContext interface, with a via to
access the package-private instance by calling
internalGetLayoutContext().

Where necessary, the layout context is cast to its internal type, when
its not clear whether the parent container is a Graph or a
GraphContainer. Because of this, the getLayoutContext() have been made
final.