Make refresh rate of DeferredUpdateManager & TargetingTool configurable
This adds a new setRefreshRate() for both classes, which defines how
often paint requests are executed/the auto-exposer helper is evaluated.

Both tasks are currently executed as fast possible, which may lead to to
a very high CPU cost with very little gain. By setting a high rate,
users may artificially throttle the speed at which those operations are
performed, at the cost of responsiveness.

Relates to:
https://github.com/eclipse/gef-classic/issues/430
https://github.com/eclipse/gef-classic/issues/492