Make refresh rate of several GEF/Draw2D classes configurable
This adds a new setRefreshRate() method to the Thumbnail,
DeferredUpdateManager and TargetingTool class which defines how
often UI events are executed.

By default, those events are executed as fast possible, which may lead
to to a very high CPU cost with very little gain. By setting a high
rate, users may artificially throttle the speed at which those
operations are performed, at the cost of responsiveness.

Relates to:
https://github.com/eclipse/gef-classic/issues/430
https://github.com/eclipse/gef-classic/issues/492