Make drawer and tool entry figures configurable via PaletteCustomizer
This adds the public interfaces IDrawerFigure and IToolEntryFigure,
which are implemented by the figures created in the DrawerEditPart and
ToolEntryEditPart, respectively.

Those interfaces can be accessed with with configure methods in the
PaletteCustomizers. Clients can use them to adjust the figures without
having to re-implement the palette editpart-factory and without having
to reference the internal classes.